Second Commit
This commit adds to views, including a view that gives the list of books for a particular author. It also includes most of the forms to add materials to the database. What is still lacking is a form to update the readings which allows the user to enter the finish date.